Fix WebConsole allows saving empty program name
